LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-shorthand-expected.txt

CONSOLE MESSAGE: line 14: ReferenceError: Can't find variable: test_shorthand_value

FAIL CSS Transitions: transition sets longhands ReferenceError: Can't find variable: test_shorthand_value
